Create alarm commands
